+++ contracts/test/StateConnector.sol
@@ -11,7 +11,7 @@ contract StateConnector {
     uint32 chainId,
     bytes32 txId,
     bytes32 destinationHash,
-    uint64 amount,
+    uint256 amount,
     bytes32 currencyHash
   )
     external view
